fix: Adjust the meaning of CSPELL_DEFAULT_CONFIG_PATH
