+++ packages/cspell/samples/environment_config/cspell.custom.json
@@ -1,7 +1,9 @@
 {
     "$schema": "https://raw.githubusercontent.com/streetsidesoftware/cspell/main/cspell.schema.json",
-    // Version of the setting file.  Always 0.1
     "version": "0.2",
-    // language - current active spelling language
-    "language": "en"
+    "language": "en",
+    "words": [
+        "everyline"
+    ],
+    "noConfigSearch": true
 }
