+++ packages/cspell/samples/environment_config/cspell.yaml
@@ -1,0 +1,7 @@
+ignorePaths:
+  - "**/cspell.*.{json,yaml,yml}"
+words:
+  - everyline
+  - includegraphics
+  - someissues
+  - Zotero
