+++ packages/cspell/src/app/application.mts
@@ -20,7 +20,8 @@ import { CSpellReporterConfiguration } from './models.js';
 import type { BaseOptions, LegacyOptions, LinterCliOptions, SuggestionOptions, TraceOptions } from './options.js';
 import { fixLegacy } from './options.js';
 import { simpleRepl } from './repl/index.js';
-import { fileInfoToDocument, readConfig, readFileInfo } from './util/fileHelper.js';
+import { readConfig } from './util/configFileHelper.js';
+import { fileInfoToDocument, readFileInfo } from './util/fileHelper.js';
 import { finalizeReporter } from './util/reporters.js';
 import { readStdin } from './util/stdin.js';
 import { getTimeMeasurer } from './util/timer.js';
